cdp-warehouse-service/requirement.txt
flask==1.1.2
flask-restx==0.2.0
python-rapidjson==1.0.0
dataclasses==0.6
validator==0.7.0
pymongo==3.11.3
python-dotenv==0.15.0
dnspython==2.1.0
flask-api==2.0
pyjwt==2.0.1
pre-commit==2.10
psycopg2==2.8.6
sqlalchemy==1.3.20
numpy==1.19.4
redshift-sqlalchemy==0.2
pandas==1.1.5
flask-swagger-ui==3.36.0